Move DHT storage repair out of chord
