+++ crates/core/src/dht/stabilization.rs
@@ -75,7 +75,7 @@ impl Stabilizer {
 
     async fn handle_storage_repair_action(&self, act: PeerRingAction) -> Result<()> {
         for delivery in act.storage_sync_deliveries()? {
-            let (_next, msg) = SyncEntriesWithSuccessor::from_delivery(delivery);
+            let msg = SyncEntriesWithSuccessor::from_delivery(delivery);
             self.transport.send_storage_sync(msg).await?;
         }
         Ok(())
