+++ crates/core/src/dht/storage/sync.rs
@@ -3,13 +3,13 @@ use std::str::FromStr;
 use async_trait::async_trait;
 use serde::Serialize;
 
-use super::PeerRing;
-use super::PeerRingAction;
 use super::StorageSyncDestination;
 use super::StorageSyncTarget;
 use crate::consts::MAX_CHUNK_ENVELOPE_OVERHEAD;
 use crate::consts::TRANSPORT_CUSTOM_OVERHEAD;
 use crate::consts::TRANSPORT_MAX_SIZE;
+use crate::dht::chord::PeerRing;
+use crate::dht::chord::PeerRingAction;
 use crate::dht::entry::Entry;
 use crate::dht::entry::PlacedEntry;
 use crate::dht::entry::SyncedEntryAck;
